pyproject.toml
[project]
name = "git-autometa"
version = "0.1.0"
description = "A tool for automating JIRA-based git workflows"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "GitPython>=3.1.0",
    "PyYAML>=6.0",
    "keyring>=23.0.0",
    "rich>=12.0.0",
]

[project.scripts]
git-autometa = "git_autometa.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]
